Added options and fixed bugs.
Added support for a configuration file and added in several options
for command line flags. Added in support to ask for missing parameters.
Fixed a bug where the server connect would fail if the server name
was invalid. Fixed a bug that occured when the file name was too
generic and would match lines that were invalid. Added in smarter
support for a destination directory for downloaded files.