Fixed bot hanging on used nick.
Github Issue: https://github.com/mac-reid/lazydcc/issues/5

Fixed the issue of nick collisions when joining arbitrary servers.

If the nick is used, we keep adding underbars, _, until we find an
unused nick.

This fix slows down the server joining procedure as it needs to talk
back and forth with the server until a suitable nick is found.